Moved core.seproxy.util elements to core.util
