+++ java/component/keyple-core/src/main/java/org/eclipse/keyple/core/util/Observable.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.keyple.core.seproxy.util;
+package org.eclipse.keyple.core.util;
 
 
 import java.util.HashSet;
