+++ java/component/keyple-core/src/test/java/org/eclipse/keyple/core/util/ObservableTest.java
@@ -11,7 +11,6 @@
  ********************************************************************************/
 package org.eclipse.keyple.core.util;
 
-import org.eclipse.keyple.core.seproxy.util.Observable;
 import org.junit.Assert;
 import org.junit.Test;
