+++ java/component/keyple-plugin/pcsc/src/test/java/org/eclipse/keyple/plugin/pcsc/PcscReaderActualTest.java
@@ -14,7 +14,7 @@ package org.eclipse.keyple.plugin.pcsc;
 import org.eclipse.keyple.core.seproxy.event.ReaderEvent;
 import org.eclipse.keyple.core.seproxy.exception.KeypleReaderException;
 import org.eclipse.keyple.core.seproxy.plugin.AbstractObservableReader;
-import org.eclipse.keyple.core.seproxy.util.Observable;
+import org.eclipse.keyple.core.util.Observable;
 import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
